feat: rename pasta of validationSchema
